fix: resolve database type mismatch errors in Rust backend
- Fix INTEGER vs BIGINT type conflicts in SQL migrations
- Cast get_active_participant_count() to bigint in queries
- Resolve i32 vs i64 type mismatches in participant/session repositories
- Update health check query to use correct return type
- Enable proper session/participant count functionality

+++ backend_rust/api-server/src/database/postgres.rs
@@ -12,7 +12,7 @@ pub async fn create_pool(config: &AppConfig) -> AppResult<PgPool> {
         .await?;
 
     // Test the connection
-    let row: (i64,) = sqlx::query_as("SELECT 1")
+    let row: (i32,) = sqlx::query_as("SELECT 1")
         .fetch_one(&pool)
         .await?;
     
@@ -26,7 +26,7 @@ pub async fn create_pool(config: &AppConfig) -> AppResult<PgPool> {
 
 /// Health check for database connection
 pub async fn health_check(pool: &PgPool) -> AppResult<()> {
-    let _: (i64,) = sqlx::query_as("SELECT 1")
+    let _: (i32,) = sqlx::query_as("SELECT 1")
         .fetch_one(pool)
         .await?;
     Ok(())
